refactor parsing function into parse.cpp/hpp
